feat: created the main component
